Extract interface from ChunkManager called ChunkManager and rename it to base.
