+++ kaldb/src/main/java/com/slack/kaldb/chunkManager/CachingChunkManager.java
@@ -19,7 +19,7 @@ import org.slf4j.LoggerFactory;
  * Chunk manager implementation that supports loading chunks from S3. All chunks are readonly, and
  * commands to operate with the chunks are made available through ZK.
  */
-public class CachingChunkManager<T> extends ChunkManager<T> {
+public class CachingChunkManager<T> extends ChunkManagerBase<T> {
   private static final Logger LOG = LoggerFactory.getLogger(CachingChunkManager.class);
 
   private final MeterRegistry meterRegistry;
