--- kaldb/src/main/java/com/slack/kaldb/chunkManager/ChunkManagerApi.java
@@ -1,11 +1,0 @@
-package com.slack.kaldb.chunkManager;
-
-import com.slack.kaldb.logstore.search.SearchQuery;
-import com.slack.kaldb.logstore.search.SearchResult;
-import java.io.IOException;
-
-public interface ChunkManagerApi<T> {
-  void addMessage(T message, long msgSize, String kafkaPartitionId, long offset) throws IOException;
-
-  SearchResult<T> query(SearchQuery query);
-}
